fix #6992: expose monaco indexed db
