+++ packages/monaco/src/browser/monaco-theming-service.ts
@@ -16,7 +16,6 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import * as idb from 'idb';
 import { injectable, inject } from 'inversify';
 import * as jsoncparser from 'jsonc-parser';
 import * as plistparser from 'fast-plist';
@@ -24,7 +23,8 @@ import { ThemeService, BuiltinThemeProvider } from '@theia/core/lib/browser/them
 import URI from '@theia/core/lib/common/uri';
 import { Disposable, DisposableCollection } from '@theia/core/lib/common/disposable';
 import { FileSystem } from '@theia/filesystem/lib/common/filesystem';
-import { MonacoThemeRegistry, ThemeMix } from './textmate/monaco-theme-registry';
+import { MonacoThemeRegistry } from './textmate/monaco-theme-registry';
+import { getThemes, putTheme, MonacoThemeState } from './monaco-indexed-db';
 
 export interface MonacoTheme {
     id?: string;
@@ -53,17 +53,6 @@ export interface MonacoThemeJson {
      * Themes can include each other. It specifies how inclusions should be resolved.
      */
     includes?: { [includePath: string]: any }
-}
-
-let monacoDB: Promise<idb.IDBPDatabase> | undefined;
-if ('indexedDB' in window) {
-    monacoDB = idb.openDB('theia-monaco', 1, {
-        upgrade: db => {
-            if (!db.objectStoreNames.contains('themes')) {
-                db.createObjectStore('themes', { keyPath: 'id' });
-            }
-        }
-    });
 }
 
 @injectable()
@@ -171,68 +160,38 @@ export class MonacoThemingService {
         this.toUpdateUiTheme.push(Disposable.create(() => document.body.classList.remove(uiTheme)));
     }
 
-    protected static doRegister(state: MonacoThemingService.MonacoThemeState): Disposable {
+    protected static doRegister(state: MonacoThemeState): Disposable {
         const { id, label, description, uiTheme, data } = state;
         const type = uiTheme === 'vs' ? 'light' : uiTheme === 'vs-dark' ? 'dark' : 'hc';
         const builtInTheme = uiTheme === 'vs' ? BuiltinThemeProvider.lightCss : BuiltinThemeProvider.darkCss;
-        const toDispose = new DisposableCollection(ThemeService.get().register({
-            type,
-            id,
-            label,
-            description: description,
-            editorTheme: data.name!,
-            activate(): void {
-                builtInTheme.use();
-            },
-            deactivate(): void {
-                builtInTheme.unuse();
-            }
-        }));
-        this.storeTheme(state, toDispose);
-        return toDispose;
+        return new DisposableCollection(
+            ThemeService.get().register({
+                type,
+                id,
+                label,
+                description: description,
+                editorTheme: data.name!,
+                activate(): void {
+                    builtInTheme.use();
+                },
+                deactivate(): void {
+                    builtInTheme.unuse();
+                }
+            }),
+            putTheme(state)
+        );
     }
 
     protected static async restore(): Promise<void> {
-        if (!monacoDB) {
-            return;
-        }
         try {
-            const db = await monacoDB;
-            const themes = await db.transaction('themes', 'readonly').objectStore('themes').getAll();
+            const themes = await getThemes();
             for (const state of themes) {
-                if (MonacoThemingService.MonacoThemeState.is(state)) {
-                    MonacoThemeRegistry.SINGLETON.setTheme(state.data.name!, state.data);
-                    MonacoThemingService.doRegister(state);
-                }
+                MonacoThemeRegistry.SINGLETON.setTheme(state.data.name!, state.data);
+                MonacoThemingService.doRegister(state);
             }
         } catch (e) {
             console.error('Failed to restore monaco themes', e);
         }
-    }
-
-    protected static async storeTheme(state: MonacoThemingService.MonacoThemeState, toDispose: DisposableCollection): Promise<void> {
-        if (!monacoDB) {
-            return;
-        }
-        const db = await monacoDB;
-        if (toDispose.disposed) {
-            return;
-        }
-        const id = state.id;
-        await db.transaction('themes', 'readwrite').objectStore('themes').put(state);
-        if (toDispose.disposed) {
-            await this.cleanTheme(id);
-            return;
-        }
-        toDispose.push(Disposable.create(() => this.cleanTheme(id)));
-    }
-
-    protected static async cleanTheme(id: string): Promise<void> {
-        if (!monacoDB) {
-            return;
-        }
-        const db = await monacoDB;
-        await db.transaction('themes', 'readwrite').objectStore('themes').delete(id);
     }
 
     /* remove all characters that are not allowed in css */
@@ -244,18 +203,4 @@ export class MonacoThemingService {
         return str;
     }
 
-}
-export namespace MonacoThemingService {
-    export interface MonacoThemeState {
-        id: string,
-        label: string,
-        description?: string,
-        uiTheme: monaco.editor.BuiltinTheme
-        data: ThemeMix
-    }
-    export namespace MonacoThemeState {
-        export function is(state: Object | undefined): state is MonacoThemeState {
-            return !!state && typeof state === 'object' && 'id' in state && 'label' in state && 'uiTheme' in state && 'data' in state;
-        }
-    }
 }
